requirements.txt
boto3
openai-whisper
